Move the stuff to icu_provider_source
